test: use require.IsType and consolidate benchmark into metafields_test.go
Replace bare type assertion + require.True(ok) patterns with require.IsType
in metafields_other_test.go and metafields_windows_test.go. Move benchmark
functions from metafields_bench_test.go into metafields_test.go and delete
the separate bench file, following the one-test-file-per-processor principle.
