+++ libbeat/reader/readfile/metafields_other_test.go
@@ -42,10 +42,10 @@ func createTestFileInfo() file.ExtendedFileInfo {
 func checkFields(t *testing.T, expected, actual mapstr.M) {
 	t.Helper()
 
-	logMap, ok := actual["log"].(mapstr.M)
-	require.True(t, ok, "expected log to be mapstr.M")
-	fileMap, ok := logMap["file"].(mapstr.M)
-	require.True(t, ok, "expected log.file to be mapstr.M")
+	require.IsType(t, mapstr.M{}, actual["log"], "expected log to be mapstr.M")
+	logMap := actual["log"].(mapstr.M)
+	require.IsType(t, mapstr.M{}, logMap["file"], "expected log.file to be mapstr.M")
+	fileMap := logMap["file"].(mapstr.M)
 
 	require.Equal(t, "17", fileMap[deviceIDKey])
 	delete(fileMap, deviceIDKey)
@@ -63,10 +63,10 @@ func checkFields(t *testing.T, expected, actual mapstr.M) {
 func checkFieldsWithOwnerGroup(t *testing.T, expected, actual mapstr.M) {
 	t.Helper()
 
-	logMap, ok := actual["log"].(mapstr.M)
-	require.True(t, ok, "expected log to be mapstr.M")
-	fileMap, ok := logMap["file"].(mapstr.M)
-	require.True(t, ok, "expected log.file to be mapstr.M")
+	require.IsType(t, mapstr.M{}, actual["log"], "expected log to be mapstr.M")
+	logMap := actual["log"].(mapstr.M)
+	require.IsType(t, mapstr.M{}, logMap["file"], "expected log.file to be mapstr.M")
+	fileMap := logMap["file"].(mapstr.M)
 
 	require.Equal(t, "17", fileMap[deviceIDKey])
 	delete(fileMap, deviceIDKey)
