+++ libbeat/reader/readfile/metafields_windows_test.go
@@ -50,10 +50,10 @@ func createTestFileInfo() file.ExtendedFileInfo {
 func checkFields(t *testing.T, expected, actual mapstr.M) {
 	t.Helper()
 
-	logMap, ok := actual["log"].(mapstr.M)
-	require.True(t, ok, "expected log to be mapstr.M")
-	fileMap, ok := logMap["file"].(mapstr.M)
-	require.True(t, ok, "expected log.file to be mapstr.M")
+	require.IsType(t, mapstr.M{}, actual["log"], "expected log to be mapstr.M")
+	logMap := actual["log"].(mapstr.M)
+	require.IsType(t, mapstr.M{}, logMap["file"], "expected log.file to be mapstr.M")
+	fileMap := logMap["file"].(mapstr.M)
 
 	require.Equal(t, "100", fileMap[idxhiKey])
 	delete(fileMap, idxhiKey)
